Moved some GML2 classes to subpackages
