+++ source/imaer-gml/src/main/java/nl/overheid/aerius/gml/base/result/GML2Result.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see http://www.gnu.org/licenses/.
  */
-package nl.overheid.aerius.gml.base;
+package nl.overheid.aerius.gml.base.result;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -22,12 +22,10 @@ import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import nl.overheid.aerius.gml.base.FeatureMember;
+import nl.overheid.aerius.gml.base.GMLConversionData;
+import nl.overheid.aerius.gml.base.IsGmlProperty;
 import nl.overheid.aerius.gml.base.geo.GML2Geometry;
-import nl.overheid.aerius.gml.base.result.IsGmlCalculationPoint;
-import nl.overheid.aerius.gml.base.result.IsGmlCustomCalculationPoint;
-import nl.overheid.aerius.gml.base.result.IsGmlNSLCalculationPoint;
-import nl.overheid.aerius.gml.base.result.IsGmlReceptorPoint;
-import nl.overheid.aerius.gml.base.result.IsGmlResult;
 import nl.overheid.aerius.shared.domain.result.EmissionResultKey;
 import nl.overheid.aerius.shared.domain.v2.geojson.Geometry;
 import nl.overheid.aerius.shared.domain.v2.geojson.Point;
